# Optimize datetime static constructor