--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- [#44](https://github.com/aeon-php/calendar/pull/46) - **Optimize datetime static constructor**  - [@norberttech](https://github.com/norberttech)
 - [#43](https://github.com/aeon-php/calendar/pull/43) - **Added more benchmarks** - [@norberttech](https://github.com/norberttech)
 - [#45](https://github.com/aeon-php/calendar/pull/45) - **new static create  method for Day and Month classes - [@eamirgh](https://github.com/eamirgh)
 - [#40](https://github.com/aeon-php/calendar/pull/40) - **created timeBetween method for better experience issue - [@eamirgh](https://github.com/eamirgh)