--- a/benchmark/Aeon/Calendar/Benchmark/Gregorian/InitializationBench.php
+++ b/benchmark/Aeon/Calendar/Benchmark/Gregorian/InitializationBench.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Aeon\Calendar\Benchmark\Gregorian;
 
-use Aeon\Calendar\Gregorian\DateTime;
 use Aeon\Calendar\Gregorian\Day;
 use Aeon\Calendar\Gregorian\Month;
 use Aeon\Calendar\Gregorian\Time;
@@ -18,11 +17,6 @@ use Aeon\Calendar\Gregorian\Year;
  */
 final class InitializationBench
 {
-    public function bench_datetime_immutable() : void
-    {
-        new \DateTimeImmutable('2020-01-01 00:00:00.00000 UTC');
-    }
-
     public function bench_aeon_year() : void
     {
         new Year(2020);
@@ -46,29 +40,5 @@ final class InitializationBench
     public function bench_aeon_time() : void
     {
         new Time(00, 00, 00, 0);
-    }
-
-    public function bench_aeon_datetime_from_datetime_immutable() : void
-    {
-        DateTime::fromDateTime(new \DateTimeImmutable('2020-01-01 00:00:00.00000 UTC'));
-    }
-
-    public function bench_aeon_datetime_create() : void
-    {
-        DateTime::create(2020, 01, 01, 00, 00, 00, 0, 'UTC');
-    }
-
-    public function bench_aeon_datetime_new() : void
-    {
-        new DateTime(
-            new Day(new Month(new Year(2020), 01), 01),
-            new Time(00, 00, 00, 0),
-            new TimeZone('UTC')
-        );
-    }
-
-    public function bench_aeon_datetime_from_string() : void
-    {
-        DateTime::fromString('2020-01-01 00:00:00.00000 UTC');
     }
 }