--- a/src/Aeon/Calendar/Gregorian/Day.php
+++ b/src/Aeon/Calendar/Gregorian/Day.php
@@ -54,7 +54,6 @@ final class Day
     }
 
     /**
-     * @throws \Exception
      * @psalm-pure
      */
     public static function fromString(string $date) : self