**Before optimization**

```
+-------------------+---------------------------------------------+------+-----+---------+------------+--------+
| benchmark         | subject                                     | revs | its | mean    | mem_peak   | diff   |
+-------------------+---------------------------------------------+------+-----+---------+------------+--------+
| DateTimeInitBench | bench_datetime_immutable                    | 1000 | 5   | 0.001ms | 5,059,040b | 1.00x  |
| DateTimeInitBench | bench_aeon_datetime_from_datetime_immutable | 1000 | 5   | 0.105ms | 6,503,384b | 80.69x |
| DateTimeInitBench | bench_aeon_datetime_create                  | 1000 | 5   | 0.048ms | 6,503,072b | 37.11x |
| DateTimeInitBench | bench_aeon_datetime_new                     | 1000 | 5   | 0.054ms | 6,503,072b | 41.50x |
| DateTimeInitBench | bench_aeon_datetime_from_string             | 1000 | 5   | 0.099ms | 6,503,384b | 76.26x |
+-------------------+---------------------------------------------+------+-----+---------+------------+--------+
```

**After optimization**

```
+-------------------+---------------------------------------------+------+-----+---------+------------+--------+
| benchmark         | subject                                     | revs | its | mean    | mem_peak   | diff   |
+-------------------+---------------------------------------------+------+-----+---------+------------+--------+
| DateTimeInitBench | bench_datetime_immutable                    | 1000 | 5   | 0.001ms | 5,059,040b | 1.00x  |
| DateTimeInitBench | bench_aeon_datetime_from_datetime_immutable | 1000 | 5   | 0.024ms | 6,367,184b | 18.52x |
| DateTimeInitBench | bench_aeon_datetime_create                  | 1000 | 5   | 0.011ms | 6,503,480b | 8.61x  |
| DateTimeInitBench | bench_aeon_datetime_new                     | 1000 | 5   | 0.011ms | 6,503,480b | 8.31x  |
| DateTimeInitBench | bench_aeon_datetime_from_string             | 1000 | 5   | 0.024ms | 6,367,184b | 18.06x |
+-------------------+---------------------------------------------+------+-----+---------+------------+--------+
```

`bench_aeon_datetime_from_datetime_immutable` is 60 times faster now.